Изменить логику, созласно замечаниям
