+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -1,73 +1,39 @@
 package ru.yandex.practicum.filmorate.storage;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.exceptions.EntityNotFoundException;
-import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.service.UserService;
 
-import java.time.LocalDate;
-import java.util.HashMap;
 import java.util.Map;
 
 @Slf4j
 @Component
 public class InMemoryUserStorage implements UserStorage {
 
-    private static final Map<Integer, User> users = new HashMap<>();
-    private static int id;
-
+    private final UserService userService;
+    @Autowired
+    public InMemoryUserStorage(UserService userService) {
+        this.userService = userService;
+    }
     @Override
     public User create(User user) {
-        validate(user);
-        if (user.getId() == 0) {
-            user.setId(++id);
-        }
-        users.put(user.getId(), user);
-        log.info("Пользователь с id '{}' добавлен в список", user.getId());
-        return user;
+        return userService.create(user);
     }
-
     @Override
     public Map<Integer, User> readAll() {
-        return users;
+        return userService.readAll();
 
     }
-
     @Override
     public User update(User user) {
-        if (users.containsKey(user.getId())) {
-            validate(user);
-            users.put(user.getId(), user);
-            log.info("Пользователь с id '{}' обновлен", user.getId());
-            return user;
-        } else {
-            log.info("NoMovieFoundException (Пользователь не может быть обновлен, т.к. его нет в списке)");
-            throw new EntityNotFoundException("Пользователь не может быть обновлен, т.к. его нет в списке");
-        }
+        return userService.update(user);
+    }
+    @Override
+    public User getUserById(int id) {
+        return userService.getUserById(id);
     }
 
-    public void validate(User user) {
-        if (user.getEmail().isEmpty() || user.getEmail().isBlank() || !(user.getEmail().contains("@"))) {
-            log.info("ValidationException (Пустой email или email не содержит '@')");
-            throw new ValidationException("Пустой email или email не содержит '@'");
-        }
-        if (user.getLogin().isEmpty() || user.getLogin().isBlank() || user.getLogin().contains(" ")) {
-            log.info("ValidationException (Пустой логин или логин содержит пробелы)");
-            throw new ValidationException("Пустой логин или логин содержит пробелы");
-        }
-        if (user.getName().isEmpty() || user.getName().isBlank()) {
-            log.info("Имя пользователя соответствует его логину");
-            user.setName(user.getLogin());
-        }
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.info("ValidationException (Дата рождения указана в будущем времени)");
-            throw new ValidationException("Дата рождения указана в будущем времени");
-        }
-        if (user.getId() < 0) {
-            log.info("ValidationException (Значение id не может быть отрицательным)");
-            throw new ValidationException("Значение id не может быть отрицательным");
-        }
-    }
 }
