+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,7 +1,5 @@
 package ru.yandex.practicum.filmorate.storage;
 
-import ru.yandex.practicum.filmorate.exceptions.EntityNotFoundException;
-import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -13,6 +11,8 @@ public interface UserStorage {
     Map<Integer, User> readAll();
 
     User update(User user);
+
+    User getUserById(int id);
